performance enhancements. UI fixes. Fork and Parallel support.
there still a lot to enhance for performance. (SPECIALLY for big images.) 
bug generally speaking the tools works correctly for 2k or less resolution
and generally speaking works fine. but: 


For big images the thumbs uses the whole image. we should use a rescale instance. 
so it renders better. 
also given we fiddle with the bytes it also affects rendering. 
for big images our approach to create the samples is good for one of 
but we request a bulk, we could enhance the approach to do less loops 
and process several channels at the time.

but for the time and needs it works fine We will put a pin into 
this changes that still needed for later revisit. 

Moving on we need to support detection of "pseudorandom noise" 
Read the image Strings (on the binary file or URL data) 

and to transform color bytes into text.