Migrated some of the code into a external module as they can be used for more functionality on more applications.
adding a new Package dependency for reading Image metadata. 
enhanced Scaling image code. to be used on the future to create copies of image
that are better performant. 

migrated code from a Single class into 3. 
from Canvas Container into
-ImageColorUtilities. 
   there is code that is more about transformation than properties of the canvas.
-WrappedImage. 
	this class will encapsulate the OG image loaded from the source. and only
allow it to be read. thus avoiding mutation and allowing us to work in multiple threads with the same image. 

some more changed are required here for better readablility. still working on